+++ test/integration/firefox_test/serverspec/default_spec.rb
@@ -4,7 +4,7 @@ require 'serverspec'
 set :backend, :exec
 
 describe 'firefox::default' do
-  describe command('firefox --version') do
+  describe command('firefox -v') do
     its(:stdout) { should match(/Mozilla Firefox /) }
     its(:exit_status) { should eq 0 }
   end