+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## 2.0.0
+* Add support for Mac OS X, CentOS, and Ubuntu platforms
 * Automatically determine latest Firefox package version
 * Add firefox_version method that returns version installed
-* Add a ability to select language 
-* Replace download url with releases uri
+* Add a ability to select language for Windows and Mac OS X platforms 
+* Replace download URL with releases URL
 
 ## 1.1.1
 * Install Firefox 32.0.3